# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVADIA API",
        "description": "API complète de la plateforme de gestion hôtelière EVADIA. Comprend l'authentification, la gestion admin, le back-office hôtel et les endpoints client.",
        "contact": {
            "name": "EVADIA",
            "email": "contact@evadia.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/admin/messages": {
            "get": {
                "tags": [
                    "Admin - Messagerie"
                ],
                "summary": "Liste des conversations",
                "description": "Retourne les conversations de l'admin regroupées par interlocuteur.",
                "operationId": "7eb85dcd181b988e3c51d6f6bc7968a0",
                "responses": {
                    "200": {
                        "description": "Liste des conversations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "interlocuteur": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "nom": {
                                                                "type": "string",
                                                                "example": "Martin"
                                                            },
                                                            "prenom": {
                                                                "type": "string",
                                                                "example": "Sophie"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "sophie@hotel.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "dernier_message": {
                                                        "type": "string",
                                                        "example": "Merci pour votre réponse"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "non_lus": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Messagerie"
                ],
                "summary": "Envoyer un message",
                "description": "Envoie un nouveau message à un utilisateur. Déclenche une notification temps réel via WebSocket.",
                "operationId": "42cd8a5b57bbd3666c9ce50f6017f965",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "destinataire_id",
                                    "contenu"
                                ],
                                "properties": {
                                    "destinataire_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "sujet": {
                                        "type": "string",
                                        "example": "Concernant votre hôtel",
                                        "nullable": true
                                    },
                                    "contenu": {
                                        "type": "string",
                                        "example": "Bonjour, votre demande a été traitée."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message envoyé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message envoyé."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/messages/conversation/{userId}": {
            "get": {
                "tags": [
                    "Admin - Messagerie"
                ],
                "summary": "Messages d'une conversation",
                "description": "Retourne tous les messages échangés avec un utilisateur. Marque les messages reçus comme lus.",
                "operationId": "9f4ec0ad2a5623bb1c83747ab7385965",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages de la conversation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "interlocuteur": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Martin"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "Sophie"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "expediteur_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "contenu": {
                                                        "type": "string",
                                                        "example": "Bonjour, comment puis-je vous aider ?"
                                                    },
                                                    "sujet": {
                                                        "type": "string",
                                                        "example": "Demande de renseignement",
                                                        "nullable": true
                                                    },
                                                    "lu": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "date_envoi": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtenu via /api/auth/login ou /api/auth/register",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}